code64_automation {LICHospitalR}	R Documentation

Code 64 Email Automation

Description

Run the code64_charged_accounts_query() and email them to Performance Improvement and the Medical Chairperson

The RDCOMClient Library must be called into the namespace first with any of the following

library(RDCOMClient)

require(RDCOMClient)

if(!require(pacman)) install.packages("pacman")

pacman::p_load("RDCOMClient")

Usage


code64_automation(.delete_file = FALSE, .email)

Arguments

.delete_file
Default is FALSE, TRUE will delete file.

.email
Provide the email address for the recipient. The email must be in double quotes like so: c("person@licommunityhospital.org;person2@licommunityhospital.org") using a semi-colon if there is more than one address.

Details

Requires a connection to DSS, uses both db_connect() and db_disconnect()

Will ask you where you want to save the file so you can email it out, will delete the file upon email send.

Uses the code64_charged_accounts_query() internally.

Author(s)

Steven P. Sanderson II, MPH

Examples


## Not run: 
library(RDCOMClient)
code64_automation(.email = "someone@email.com")

## End(Not run)


[Package LICHospitalR version 0.2.0 Index]