coded_consults_query {LICHospitalR}	R Documentation

OPPE Coded Consults Query

Description

Get the coded consults for a specified provider using their ID number or resp_pty cd.

Usage


coded_consults_query(.resp_pty)

Arguments

.resp_pty
The providers six digit ID Number. This should be quoted like '123456'

Details

Get the coded consults for a specified consulting provider using their ID number/resp_pty cd.

This gets data from DSS, you must have a connection file on your pc to use this function. If you do not have one then submit a help desk ticket for one.

Gets data from smsdss.c_Coded_Consults_v for the last 6 full months by discharge date.

Uses timetk::tk_augment_timeseries_signature() to obtain a few different time components

Value

A tibble

Author(s)

Steven P. Sanderson II, MPH

Examples


library(janitor)
library(dplyr)
library(tibble)
library(DBI)
library(odbc)
library(lubridate)
library(timetk)
library(tidyquant)
coded_consults_query(.resp_pty = '123456')


[Package LICHospitalR version 0.2.0 Index]