coded_consults_seasonal_diagnositcs_plt {LICHospitalR}	R Documentation

OPPE Coded Consults Seasonal Diagnostics

Description

Get the seasonal diagnostics of the last 6 months consulant usage

Usage


coded_consults_seasonal_diagnositcs_plt(.data, .date_col, .value_col)

Arguments

.data
The data you want to analyze

.date_col
The column that has the date you want to use, typically dsch_date

.value_col
The column that holds the value, can use record_flag from the query

Details

Must have data supplied either by import or by coded_consults_query() function

Must have a date column (adm_date or dsch_date)

Must have a value column (should just use record_flag from the query)

Value

A plotly plot

Author(s)

Steven P. Sanderson II, MPH

Examples


library(timetk)
library(dplyr)
library(tibble)
coded_consults_seasonal_diagnositcs_plt(
  .data = coded_consults_query(.resp_pty = "013128")
  , .date_col = dsch_date
  , .value_col = record_flag
  )


[Package LICHospitalR version 0.2.0 Index]