coded_consults_top_providers_tbl {LICHospitalR}	R Documentation

OPPE Coded Consults Top Attending

Description

Gets the top n attending providers who have consulted the consultant in question.

Usage


coded_consults_top_providers_tbl(
  .data,
  .top_n,
  .attending_col,
  .consultant_col
)

Arguments

.data
The data that you provided. There must be an Attending column and Consultant column

.top_n
How many of the top attending providers do you want returned.

.attending_col
The column that holds the name of the attending provider

.consultant_col
The column that holds the name of the consulting provider(s)

Details

Get the top n attending providers and counts of times they consulted the specialist provider in question.

This gets data from either DSS by using the coded_consults_query() or an imported file.

Author(s)

Steven P. Sanderson II, MPH

Examples


library(dplyr)
library(tibble)
coded_consults_top_providers_tbl(
.data = coded_consults_query(.resp_pty = "013128")
, .top_n = 10
, .attending_col = attending_md
, .consultant_col = consultant
)


[Package LICHospitalR version 0.2.0 Index]