congenital_malformation_automation {LICHospitalR}	R Documentation

Congenital Malformation Automation for HIM

Description

This automation will send information on congenital malformations that comes from the congenital_malformation_query() to HIM

The RDCOMClient Library must be called into the namespace first with any of the following

library(RDCOMClient)

require(RDCOMClient)

if(!require(pacman)) install.packages("pacman")

pacman::p_load("RDCOMClient")

Usage


congenital_malformation_automation(.delete_file = FALSE, .email)

Arguments

.delete_file
Default is FALSE, TRUE will delete file

.email
Provide the email address for the recipient. The email must be in double quotes like so: c("person@licommunityhospital.org;person2@licommunityhospital.org") using a semi-colon if there is more than one address.

Details

Uses the congenital_malformation_query() internally.

Requires a connection to DSS and uses db_connect() and db_disconnect()

Author(s)

Steven P. Sanderson II, MPH

Examples


## Not run: 
library(RDCOMClient)
congenital_malformation_automation(.email= "someone@email.com")

## End(Not run)


[Package LICHospitalR version 0.2.0 Index]