denials_admits_by_ed_query {LICHospitalR}	R Documentation

Denials Admits by ED

Description

This query gets the admit counts by ed provider

Usage


denials_admits_by_ed_query(.name)

Arguments

.name
The name of the provider you want returned

Details

Queries DSS the linked WellSoft Report DB

Uses the db_connect() and db_disconnect() functions

Does not need an argument but one can be passes til filter on name if desired, although this should be done as a dplyr::filter() so this funcationality may be taken away in the future

Value

A tibble

Author(s)

Steven P. Sanderson II, MPH

Examples


denials_admits_by_ed_query()


[Package LICHospitalR version 0.2.0 Index]