denials_admits_by_md_query {LICHospitalR}	R Documentation

Denials Admits by MD

Description

This query gets the admit counts by attending provider

Usage


denials_admits_by_md_query(.name)

Arguments

.name
The name of the provider you want returned, leave empty for all

Details

Queries DSS

Uses the db_connect() and db_disconnect() functions

Does not need an argument but one can be passed to filter on name if desired, although this should be done as a dplyr::filter() so this funcationality may be taken away in the future

The data is grouped by:

Payer_Category

Attending Dr

Admit Month

Admit Year

Primary Payer Plan Code

Med Staff Department

Value

A Tibble

Author(s)

Steven P. Sanderson II, MPH

Examples


library(DBI)
library(dplyr)
library(tibble)
library(janitor)
library(data.table)

denials_admits_by_md_query()


[Package LICHospitalR version 0.2.0 Index]