geocode_discharges_query {LICHospitalR}	R Documentation

Get Address for discharges that can be geocoded

Description

Get discharged accounts from DSS

Usage


geocode_discharges_query()

Details

Requires a connection to DSS, uses both db_connect() and db_disconnect()

This will look back at discharges starting with a discharge date of six months prior to the SQL GETEDATE() function

Value

A tibble

Author(s)

Steven P. Sanderson II, MPH

Examples


library(DBI)
library(tibble)

geocode_discharges_query()


[Package LICHospitalR version 0.2.0 Index]