infection_prevention_patient_days_automation {LICHospitalR}	R Documentation

Patient Days for Infection Prevention Automation

Description

This takes in the data from infection_prevention_patient_days_query() and infection_prevention_patient_days_tbl() puts them in a single excel workbook on different sheets and sends.

The RDCOMClient Library must be called into the namespace first with any of the following

library(RDCOMClient)

require(RDCOMClient)

if(!require(pacman)) install.packages("pacman")

pacman::p_load("RDCOMClient")

Usage


infection_prevention_patient_days_automation(
  .data,
  .delete_file = FALSE,
  .email
)

Arguments

.data
The output from the tbl function infection_prevention_patient_days_tbl()

.delete_file
Default is FALSE, TRUE will delete file

.email
Provide the email address for the recipient. The email must be in double quotes like so: \emailc("person@licommunityhospital.org;person2@licommunityhospital.org") using a semi-colon if there is more than one address.

Details

Takes the output from infection_prevention_patient_days_tbl() and saves it to an excel workbook using the xlsx::saveWorkbook() function. The file is saved to a location specified by the user and upon successful function completion the file is deleted.

Author(s)

Steven P. Sanderson II, MPH

Examples


## Not run: 
library(RDCOMClient)
library(tidyverse)
infection_prevention_patient_days_query() %>%
  infection_prevention_patient_days_tbl() %>%
  infection_prevention_patient_days_automation(.email = "someone@email.com")

## End(Not run)


[Package LICHospitalR version 0.2.0 Index]