infection_prevention_patient_days_query {LICHospitalR}	R Documentation

Patient Days for Infection Prevention Query

Description

This gets the data necessary for infection prevention. This is used for a variety of reasons, one being total vent days against total days.

Usage


infection_prevention_patient_days_query()

Details

Requires a connection to DSS, uses both db_connect() and db_disconnect()

Has Start and End dates set dynamically in the query in order to obtain data for the previous month.

Value

A tibble

Author(s)

Steven P. Sanderson II, MPH

Examples


## Not run: 
infection_prevention_patient_days_query()

## End(Not run)


[Package LICHospitalR version 0.2.0 Index]