inpatient_coding_lag_tbl {LICHospitalR}	R Documentation

Inpatient Coding Lag tibble

Description

This function takes in the query inpatient_coding_lag_query() as the data argument and augments the data so that it can be taken into the inpatient_coding_lag_automation() function.

Usage


inpatient_coding_lag_tbl(.data)

Arguments

.data
Data that you want to manipulate typically from inpatient_coding_lag_query()

Details

Takes in inpatient_coding_lag_query() as the data argument.

Returns a tibble that can be used in the inpatient_coding_lag_automation() function.

Value

A tibble

Author(s)

Steven P. Sanderson II, MPH

Examples


## Not run: 
library(tidyverse)

inpatient_coding_lag_query() %>%
  inpatient_coding_lag_tbl()

## End(Not run)


[Package LICHospitalR version 0.2.0 Index]