kmeans_mapped_tbl {LICHospitalR}	R Documentation

OPPE CPOE K-Means Mapper

Description

Create a tibble that maps the oppe_cpoe_kmeans() using [purrr::map() to create a nested data.frame/tibble that holds n centers. This tibble will be used to help create a scree plot.

Usage


kmeans_mapped_tbl(.data, .centers = 15)

Arguments

.data
You must have a tibble in the working environment that is produced by using the following work flow: oppe_cpoe_query() %>% oppe_cpoe_tbl() %>% oppe_cpoe_user_item_tbl()

.centers
How many different centers do you want to try

Details

Takes in a single parameter of .centers. This is used to create the tibble and map the oppe_cpoe_kmeans() function down the list creating a nested tibble.

Value

A nested tibble

Author(s)

Steven P. Sanderson II, MPH

See Also

https://en.wikipedia.org/wiki/Scree_plot

Examples


ui_tbl <- oppe_cpoe_query() %>%
  oppe_cpoe_tbl() %>%
  oppe_cpoe_user_item_tbl()

kmeans_mapped_tbl(ui_tbl)


[Package LICHospitalR version 0.2.0 Index]