los_ra_index_plt {LICHospitalR}	R Documentation

Plot LOS and Readmit Index with Variance

Description

Plot the index of the length of stay and readmit rate against each other along with the variance

Usage


los_ra_index_plt(.data)

Arguments

.data
The data supplied from los_ra_index_summary_tbl()

Details

Expects a tibble

Expects a Length of Stay and Readmit column, must be numeric

Uses patchwork to stack plots

Value

A patchwork ggplot2 plot

Author(s)

Steven P. Sanderson II, MPH

Examples


library(dplyr)
library(ggplot2)
library(patchwork)
library(tidyquant)

los_ra_index_query() %>%
los_ra_index_summary_tbl() %>%
los_ra_index_plt()


[Package LICHospitalR version 0.2.0 Index]