los_ra_index_summary_tbl {LICHospitalR}	R Documentation

Make LOS and Readmit Index Summary Tibble

Description

Create the length of stay and readmit index summary tibble

Usage


los_ra_index_summary_tbl(.data)

Arguments

.data
The data you are going to analyze from los_ra_index_query()

Details

Expects a tibble

Adds a year column from the Admit Date and one from Discharge Date

Uses all data to compute variance, if you want it for a particular time frame you will have to filter the data from los_ra_index_query()

Value

A tibble

Author(s)

Steven P. Sanderson II, MPH

Examples


los_ra_index_summary_tbl(
  .data = los_ra_index_query()
)


[Package LICHospitalR version 0.2.0 Index]