monthly_discharge_trauma_query {LICHospitalR}	R Documentation

Monthly Discharge Trauma File Query

Description

This query is used to grab all possible Trauma related accounts. There are two queries that are run, one for accounts by admit_date and one for accounts by discharge_date.

There are no parameters to this function as of yet, it may be introduced in the future if there is a strong need for it. The dates are set dynamically in the SQL so that data is always run for the previous third month in which the query is run. For example if the query is run in any day of November 2020 then the data will be pulled for August of 2020

Usage


monthly_discharge_trauma_query()

Details

Need a valid DSS connection and rights to query.

Utilizes both db_connect() and db_disconnect() functions

Returns data as a tibble

Value

A tibble

Author(s)

Steven P. Sanderson II, MPH

Examples


## Not run: 
monthly_discharge_trauma_query()

## End(Not run)


[Package LICHospitalR version 0.2.0 Index]