monthly_psy_admits_discharges_tbl {LICHospitalR}	R Documentation

Monthly PSY Admit/Discharge tibble

Description

Takes data from monthly_psy_admits_query() and monthly_psy_discharges_query() and puts them inside of an excel workbook with data on different sheets.

There are no parameters to this function, it takes in it directly the queries internally.

The monthly_psy_admits_discharges_automation() will internally handle this.

Usage


monthly_psy_admits_discharges_tbl()

Details

Creates a sheet for the raw admit and discharge data

Creates a sheet for distinct Admitted and Discharged MRN's

Creates a sheet with pivot tables for Admited/Discharged financial grouping (i.e. Medicare, Blue Cross)

Creates a sheet a count of distinct Admitted and Discharged MRN's

Value

A java pointer to an excel workbook (or a list in the future)

Author(s)

Steven P. Sanderson II, MPH

Examples


## Not run: 
monthly_psy_admits_discharges_tbl()

## End(Not run)


[Package LICHospitalR version 0.2.0 Index]