monthly_psy_discharges_query {LICHospitalR}	R Documentation

Monthly PSY Discharges Query

Description

This query is used to grab all hospital service PSY discharges for the previous month. There are no parameters to this function as of yet, it may be introduce in the future if there is a strong need for it.

This information is used to send to Elizabeth Saporito

Usage


monthly_psy_discharges_query()

Details

Need a valid DSS connection and rights to query.

Utilizes both db_connect() and db_disconnect() functions

Returns data as a tibble

Value

A tibble

Author(s)

Steven P. Sanderson II, MPH

Examples


## Not run: 
monthly_psy_discharges_query()

## End(Not run)


[Package LICHospitalR version 0.2.0 Index]