monthly_trauma_tbl {LICHospitalR}	R Documentation

Monthly Trauma Tibble

Description

Takes data from the queries

monthly_admit_trauma_query(), and

monthly_discharge_trauma_query() And then performs the necessary transformations on them to get them into one excel file with two sheets. One for Admits and one for Discharges.

Usage


monthly_trauma_tbl()

Details

Internally gets data from monthly_admit_trauma_query()

Internally gets data from monthly_discharge_trauma_query()

Transforms them into one excel file with two sheets that can be written out to disk

Value

A java pointer to an excel workbook (or a list in the future)

Author(s)

Steven P. Sanderson II, MPH

Examples


## Not run: 
monthly_trauma_tbl()

## End(Not run)


[Package LICHospitalR version 0.2.0 Index]