myhealth_monthly_surgery_query {LICHospitalR}	R Documentation

MyHealth Monthly Surgery File

Description

This query will generate the data needed to produce the monthly MyHealth Surgery file and send to MyHealth.

Usage


myhealth_monthly_surgery_query()

Details

Need a valid DSS connection and rights to query.

Utilizes both db_connect() and db_disconnect() functions

Returns data as a tibble

This query contains the following fields:

Facilty Account Number

Provider_ID

Provider Short Name

Room ID

Start Date

Enter Procedure Room Time

Leave Procedure Room Time

Procedure Description (from ORSOS)

Anesthesia Start Date

Anesthesia Start Time

Anesthesia Stop Date

Anesthesia Stop Time

Patient Type

Admit Recovery Date

Admit Recovery Time

Leave Recovery Date

Leave Recovery Time

The date fields are dynamically set to look back one month. There are no parameters to this function

Value

A tibble

Author(s)

Steven P. Sanderson II, MPH

Examples


## Not run: 
  myhealth_monthly_surgery_query()

## End(Not run)


[Package LICHospitalR version 0.2.0 Index]