oppe_alos_plt {LICHospitalR}	R Documentation

OPPE ALOS Plots

Description

Get plots for a single providers length of stay data.

Usage


oppe_alos_plt(.data, .date_col, .by_time = "month")

Arguments

.data
The data that you would pass from oppe_alos_tbl()

.date_col
The column holding the date value

.by_time
How you want the data time aggregated, the default is "month"

Details

Takes data in from the oppe_alos_tbl() function.

Value

A patchwork time series plot

Author(s)

Steven P. Sanderson II, MPH

Examples


oppe_alos_query() %>%
  oppe_alos_tbl(.provider_id = "017236") %>%
  oppe_alos_plt(.date_col = dsch_date)


[Package LICHospitalR version 0.2.0 Index]