oppe_alos_query {LICHospitalR}	R Documentation

OPPE ALOS Query

Description

Get the average length of stay data from DSS in order to run the OPPE report. The data in this report goes back 18 months and the dates are set dynamically in the sql.

Usage


oppe_alos_query()

Details

Requires a connection to DSS, uses both db_connect() and db_disconnect()

Has start and end dates set dynamically to look back on data to ensure it has gone through the coding process.

The fields that come back are:

pt_id

ptno_num

dsch_date

dsch_yr

dsch_month

atn_dr_no

atn_dr_name

drg_no

lihn_service_line

hosim

apr_drg

severity_of_illness

los

elos

threshold

outlier_flag

drg_cost_weight

pyr_grouping

case_var

case_index

index_threshold

z_score

med_staff_dept

ward_cd

The tables that are used are:

smsdss.c_LIHN_Svc_Line_tbl

smsdss.BMH_PLM_PtAcct_V

Customer.Custom_DRG

smsdss.c_LIHN_SPARCS_BenchmarkRates

smsdss.pract_dim_v

smsdss.c_LIHN_APR_DRG_OutlierThresholds

smsdss.pyr_dim_v

smsmir.vst_rpt

Value

A tibble object

Author(s)

Steven P. Sanderson II, MPH

Examples


library(dplyr)

oppe_alos_query() %>%
glimpse()


[Package LICHospitalR version 0.2.0 Index]