oppe_alos_tbl {LICHospitalR}	R Documentation

OPPE ALOS Tibble

Description

Get the OPPE ALOS data for a provider if specified.

Usage


oppe_alos_tbl(.data, .provider_id, .provider_name)

Arguments

.data
The data to be passed into the tibble function, should be the results from oppe_alos_query()

.provider_id
The provider id number you are looking for

.provider_name
If you do not know the providers id, then you can use their name to get a match. You can also use the pract_dim_v_query() to find the id number before using this function, or you can use it inside of the id number parameter. Be careful with this approach as providers can have the same last name. It is better to use the function on it's own and select the appropriate id number. Uses stringr::str_to_lower() so that you can use all lower case letters to describe the name you are looking for.

Details

Requires a connection to DSS, uses both db_connect() and db_disconnect()

Returns the provider id and name, actual los, expected los and excess

Value

A tibble

Author(s)

Steven P. Sanderson II, MPH

Examples


oppe_alos_query() %>%
  oppe_alos_tbl()

oppe_alos_query() %>%
  oppe_alos_tbl(.provider_id = "017236")

oppe_alos_query() %>%
  oppe_alos_tbl(.provider_name = "ashraf ahad")


[Package LICHospitalR version 0.2.0 Index]