oppe_cpoe_cluster_trends_tbl {LICHospitalR}	R Documentation

OPPE CPOE Cluster Trends Table

Description

This function takes the data from the oppe_cpoe_trend_tbl() function and creates from that data along with the umap_kmeans_cluster_results_tbl from the oppe_cpoe_umap() function.

Usage


oppe_cpoe_cluster_trends_tbl(.data, .umap_data, ...)

Arguments

.data
The data that needs to be passed to the function from the oppe_cpoe_trend_tbl() function

.umap_data
The umap_kmeans_cluster_results_tbl from the oppe_cpoe_umap() function

...
Any extra columns you want passed to the function. By default the following are selected:

.cluster

order_type

You can choose the from the following as extra:

hospitalist_np_pa_flag

spclty_desc

Details

Requires data from the oppe_cpoe_trend_tbl() and oppe_cpoe_umap() functions. This data is later used to show how a particular provider stacks up against others in the same cluster.

Value

A tibble

Author(s)

Steven P. Sanderson II, MPH

Examples


q <- oppe_cpoe_query() %>%
  oppe_cpoe_tbl()

trend_tbl <- q %>%
  oppe_cpoe_trend_tbl()

ui_tbl <- q %>%
  oppe_cpoe_user_item_tbl()

kmm_tbl <- ui_tbl %>%
  kmeans_mapped_tbl()

umap_obj <- oppe_cpoe_umap(
  .data = ui_tbl
  , .kmeans_map_tbl = kmm_tbl
  , .k_cluster = 3
)

oppe_cpoe_cluster_trends_tbl(trend_tbl, umap_obj)
oppe_cpoe_cluster_trends_tbl(trend_tbl, umap_obj, hospitalist_np_pa_flag)


[Package LICHospitalR version 0.2.0 Index]