oppe_cpoe_query {LICHospitalR}	R Documentation

OPPE CPOE Query

Description

Get the CPOE data from DSS in order to run the OPPE report.

Usage


oppe_cpoe_query(.start_date = "")

Arguments

.start_date
The start date of the data, if null then it will bring back all data from the table. Uses the timetk::filter_by_time() function to filter the time.

Details

Requires a connection to DSS, uses both db_connect() and db_disconnect()

Has start and end dates set dynamically to look back on data to ensure it has gone through the coding process.

The fields that come back are:

ent_ate

req_pty_cd

pract_rpt_name

spclty_desc

hospitalist_np_pa_flag

ord_type_abbr

specimen_collect

written

verbal_order

communication

specimen_redraw

cpoe

nursing_order

The tables that are used are:

smsdss.c_CPOE_Rpt_Tbl_Rollup_v

smsdss.pract_dim_v

Value

A tibble object

Author(s)

Steven P. Sanderson II, MPH

Examples


library(dplyr)

oppe_cpoe_query() %>%
glimpse()


[Package LICHospitalR version 0.2.0 Index]