oppe_cpoe_umap_plt {LICHospitalR}	R Documentation

OPPE CPOE K-Means UMAP Projection Plot

Description

Using the uwot library we visualize the UMAP: Uniform Manifold Approximation and Projection for Dimension Reduction. https://arxiv.org/abs/1802.03426

UMAP (Uniform Manifold Approximation and Projection) is a novel manifold learning technique for dimension reduction. UMAP is constructed from a theoretical framework based in Riemannian geometry and algebraic topology. The result is a practical scalable algorithm that applies to real world data. The UMAP algorithm is competitive with t-SNE for visualization quality, and arguably preserves more of the global structure with superior run time performance. Furthermore, UMAP has no computational restrictions on embedding dimension, making it viable as a general purpose dimension reduction technique for machine learning.

Usage


oppe_cpoe_umap_plt(.data, .point_size = 2)

Arguments

.data
Takes data from the oppe_cpoe_umap() function. You simply pass in the list object and it will automatically take the cluster results tibble

.point_size
The size of the ggplot2::geom_point() size.

Details

Requires the user item table/matrix that is output from the oppe_cpoe_user_item_tbl() function.

Author(s)

Steven P. Sanderson II, MPH

Examples


ui_tbl <- oppe_cpoe_query() %>%
  oppe_cpoe_tbl() %>%
  oppe_cpoe_user_item_tbl()

km_map_tbl <- kmeans_mapped_tbl(ui_tbl)
kmeans_scree_plt(km_map_tbl)

umap_obj <- oppe_cpoe_umap(ui_tbl, km_map_tbl, 3)

oppe_cpoe_umap_plt(umap_obj)


[Package LICHospitalR version 0.2.0 Index]