oppe_cpoe_user_item_tbl {LICHospitalR}	R Documentation

OPPE CPOE K-Means

Description

Takes data from the oppe_cpoe_tbl() and transforms it into an aggregated/normalized user-item tibble of order proportions.

Usage


oppe_cpoe_user_item_tbl(.data)

Arguments

.data
The data that is passed from oppe_cpoe_tbl()

Details

This function requires that the oppe_cpoe_query() and oppe_cpoe_tbl() functions be run and takes the output of that work flow as its input. This function should be used before using a k-mean model. This is commonly referred to as a user_item matrix because "users" (e.g. providers) tend to be on the rows and "tems" (e.g. orders) on the columns.

At present this function only selects the following types of orders:

Written

Verbal

Telephone, and

CPOE

Value

A aggregated/normalized user item tibble

Author(s)

Steven P. Sanderson II, MPH

Examples


oppe_cpoe_query() %>%
  oppe_cpoe_tbl() %>%
  oppe_cpoe_user_item_tbl()


[Package LICHospitalR version 0.2.0 Index]