oppe_denials_detail_query {LICHospitalR}	R Documentation

OPPE Denials Detail Query

Description

Get the Denials Detail data for a specified provider.

Usage


oppe_denials_detail_query(.provider_id)

Arguments

.provider_id
The id of the provider you want denial detial for

Details

Requires a connection to DSS, uses both db_connect() and db_disconnect()

Has start and end dates set dynamically to look back on data to ensure it has gone through the coding process.

The fields that come back are: *

The tables that are used are:

smsdss.bmh_plm_ptacct_v

smsdss.pract_dim_v

smsdss.drg_dim_v

smsdss.c_lihn_svc_line_tbl

smsdss.c_lihn_op_svc_line_tbl

BMH-3MHIS-DB.MMM_COR_BMH_LIVE.dbo.visit_view

BMH-3MHIS-DB.MMM_COR_BMH_LIVE.dbo.CTC_VISIT

BMH-3MHIS-DB.MMM_COR_BMH_LIVE.dbo.CTC_INSURANCE

BMH-3MHIS-DB.MMM_COR_BMH_LIVE.dbo.CTC_UM_Denial

BMH-3MHIS-DB.MMM_COR_BMH_LIVE.dbo.CTC_UM_APPEAL

Value

A tibble object

Author(s)

Steven P. Sanderson II, MPH

Examples


## Not run: 
library(dplyr)

oppe_denials_detail_query(.provider_id = "005892") %>%
glimpse()

# If provider id is not known you can do either of the following
# pract_dim_v_query(.name = "the providers name") and the pract_no will be
# returned to you, or you could do the following:

oppe_denials_detail_query(
  .provider_id = pract_dim_v_query(.name = "name_here") %>%
    dplyr::pull(pract_no)
)

## End(Not run)


[Package LICHospitalR version 0.2.0 Index]