oppe_readmit_query {LICHospitalR}	R Documentation

OPPE Readmit Query

Description

Get the readmit rate data from DSS in order to run the OPPE report. The data in this report goes back 18 months and the dates are set dynamically in the sql.

Usage


oppe_readmit_query()

Details

Requires a connection to DSS, uses both db_connect() and db_disconnect()

Has start and end dates set dynamically to look back on data to ensure it has gone through the coding process.

The fields that come back are:

med_rec_no

ptno_num

adm_date

dsch_date

payor_category

atn_dr_no

med_staff_dept

lihn_svc_line

severity_of_illness

dsch_yr

dsch_qtr

dsch_month

dsch_week

dsch_day

dsch_day_name

rpt_month

rpt_qtr

dsch_disp

dsch_disp_desc

drg_no

drg_cost_weight

hospitalist_private

hospitalist_private_flag

los

interim

pt_count

readmit_count

bench_yr

readmit_rate_bench

ward_cd

z_score

pract_rpt_name

The tables that are used are:

smsdss.c_readmit_dashboard_detail_tbl

smsdss.c_readmit_dashboard_bench_tbl

smsmir.vst_rpt

Value

A tibble object

Author(s)

Steven P. Sanderson II, MPH

Examples


library(dplyr)

oppe_readmit_query() %>%
glimpse()


[Package LICHospitalR version 0.2.0 Index]