opt_bin {LICHospitalR}	R Documentation

Get the optimal bin width for a histogram given a data set, its value and the desired amount of bins

Description

Gives the optimal bindwidth for a histogram given a data set, it's value and the desired amoutn of bins

Usage


opt_bin(.data, .value_col, .iters = 30)

Arguments

.data
The data set in question

.value_col
The column that holds the values

.iters
How many times the cost function loop should run

Details

Modified from Hideaki Shimazaki Department of Physics, Kyoto University shimazaki at ton.scphys.kyoto-u.ac.jp Feel free to modify/distribute this program.

Supply a data.frame/tibble with a value column. from this an optimal binwidth will be computed for the amount of binds desired

Value

A tibble of histogram breakpoints

Author(s)

Steven P. Sanderson II, MPH

Examples


library(dplyr)
library(tibble)
library(purrr)

df_tbl <- rnorm(n = 1000, mean = 0, sd = 1)
df_tbl <- df_tbl %>%
  as_tibble() %>%
  set_names("value")

df_tbl %>%
  opt_bin(
    .value_col = value
    , .iters = 100
  )


[Package LICHospitalR version 0.2.0 Index]