orsos_j_accounts_automation {LICHospitalR}	R Documentation

ORSOS J Accounts Automation

Description

This takes in data directly from the orsos_j_accounts_query() as the data argument.

The RDCOMClient Library must be called into the namespace first with any of the following

library(RDCOMClient)

require(RDCOMClient)

if(!require(pacman)) install.packages("pacman")

pacman::p_load("RDCOMClient")

Usage


orsos_j_accounts_automation(.data, .delete_file = FALSE, .email)

Arguments

.data
The data from orsos_j_accounts_query()

.delete_file
FALSE is the default, so the file will be kept, TRUE will delete the file after function completion

.email
Provide the email address for the recipient. The email must be in double quotes like so: c("person@licommunityhospital.org;person2@licommunityhospital.org") using a semi-colon if there is more than one address.

Details

Takes in the results of orsos_j_accounts_query() function

Asks user where to save the temporary file

Will delete file after sending if parameter is set to true

Author(s)

Steven P. Sanderson II, MPH

Examples


## Not run: 
library(RDCOMClient)
library(tidyverse)
orsos_j_accounts_query() %>%
  orsos_j_accounts_automation(.email = "somone@email.com")

## End(Not run)


[Package LICHospitalR version 0.2.0 Index]