orsos_to_sproc_query {LICHospitalR}	R Documentation

ORSOS to SPROC Case Reconcilliation Query

Description

This function gets a list of accounts from smsdss.c_ORSOS_Post_Case_Rpt_Tbl and then looks for the encounter numbers in smsmir.sproc to see if the case was coded to a provider by HIM.

The start and end dates are set dynamically in order to ensure there is a two week lag in the looking for cases so that they get through the HIM coding process.

For example if we are running the report on November 5th, 2020 a Thursday, then the report will pull records from 2020-10-11 00:00:00 through 2020-10-17 23:59:00

Usage


orsos_to_sproc_query()

Details

Requires a connection to DSS, uses both db_connect() and db_disconnect()

Has start and end dates set dynamically to look back on data to ensure it has gone through the coding process.

Value

A tibble

Author(s)

Steven P. Sanderson II, MPH

Examples


orsos_to_sproc_query()


[Package LICHospitalR version 0.2.0 Index]