qec_cdi_automation {LICHospitalR}	R Documentation

CDI QEC Numbers

Description

Gather Admit / Discharges counts for CDI QEC dashboard

Usage


qec_cdi_automation(.delete_file = FALSE, .email)

Arguments

.delete_file
Default is FALSE, TRUE will delete the file.

.email
Provide the email address for the recipient. The email must be in double quotes like so: c("person@licommunityhospital.org;person2@licommunityhospital.org") using a semi-colon if there is more than one address.

Details

Requires a connection to DSS, uses both db_connect() and db_disconnect()

Will ask you where you want to save the file so you can email it out

Uses the qec_cdi_query() to automatically set the .data argument internally

Email currently goes to PMcKenna at LICommunityHospital.org, should this change please call Steven Sanderson at 2995 to update the email, and redistribute package.

Outlook may ask you to Allow the message to send

Author(s)

Steven P. Sanderson II, MPH

Examples


## Not run: 
library(RDCOMClient)
library(tidyverse)
qec_cdi_automation(.email = "someone@email.com")

## End(Not run)


[Package LICHospitalR version 0.2.0 Index]