qec_cdi_query {LICHospitalR}	R Documentation

Get CDI QEC Numbers from DSS

Description

Gets the QEC numbers for CDI

Usage


qec_cdi_query()

Details

Requires a connection to DSS, uses both db_connect() and db_disconnect()

Has Start and End dates set dynamically in the query in order to obtain for the previous month. For example if you run the query on any day in October you will get data for September.

Value

A tibble

Author(s)

Steven P. Sanderson II, MPH

Examples


library(DBI)
library(odbc)
qec_cdi_query()


[Package LICHospitalR version 0.2.0 Index]