readmit_psy_to_psy_tbl {LICHospitalR}	R Documentation

Readmit Psych To Psych Tibble

Description

This function will take the results of the readmit_psy_to_psy_query() and ensure that the results are in a tibble

Usage


readmit_psy_to_psy_tbl(.data)

Arguments

.data
The data you want to pass, namely readmit_psy_to_psy_query()

Details

Takes in the data from readmit_psy_to_psy_query()

Value

A tibble

Author(s)

Steven P. Sanderson II, MPH

Examples


## Not run: 
readmit_psy_to_psy_query() %>%
  readmit_psy_to_psy_tbl()

## End(Not run)


[Package LICHospitalR version 0.2.0 Index]