respiratory_vae_automation {LICHospitalR}	R Documentation

Respiratory VAE Automation

Description

This will email out the Respiratory VAE file to the email specified in the function. It is expected that the .data parameter be filled in with the results of the respiratory_vae_tbl() function.

The RDCOMClient Library must be called into the namespace first with any of the following

library(RDCOMClient)

require(RDCOMClient)

if(!require(pacman)) install.packages("pacman")

pacman::p_load("RDCOMClient")

Usage


respiratory_vae_automation(.data, .delete_file = FALSE, .email)

Arguments

.data
The output from the tbl function respiratory_vae_tbl()

.delete_file
Default is FALSE, TRUE will delete file

.email
Provide the email address for the recipient. The email must be in double quotes like so: c("person@licommunityhospital.org;person2@licommunityhospital.org") using a semi-colon if there is more than one address.

Details

Takes the output from respiratory_vae_tbl() and saves it to an excel workbook using the writexl::write_xlsx() function. The file is saved to a location specified by the user and upon successful function completion the file is deleted.

Author(s)

Steven P. Sanderson II, MPH

Examples


## Not run: 
library(RDCOMClient)
respiratory_vae_query() %>%
  respiratory_vae_tbl() %>%
  respiratory_vae_automation()

## End(Not run)


[Package LICHospitalR version 0.2.0 Index]