respiratory_vae_query {LICHospitalR}	R Documentation

Respiratory VAE Query

Description

This query gets the data necessary to perform the VAE (Ventilator Associated Events) calculation for the FiO2 and PEEP values for respiratory therapy.

Usage


respiratory_vae_query()

Details

Requires a connection to DSS, uses both db_connect() and db_disconnect()

Has Start and End dates set dynamically in the query in order to obtain data for the previous 7 days.

Value

A tibble

Author(s)

Steven P. Sanderson II, MPH

See Also

Infromational Link: https://www.ahrq.gov/hai/tools/mvp/modules/vae/tool.html

AHRQ Calculator (Javascript must be enabled): https://nhsn.cdc.gov/VAECalculator/vaecalc_v7.html

Examples


## Not run: 
respiratory_vae_query()

## End(Not run)


[Package LICHospitalR version 0.2.0 Index]