respiratory_vae_tbl {LICHospitalR}	R Documentation

Respiratory VAE Tibble

Description

This takes data from the respiratory_vae_query() and performs the necessary calculations to see if a possible VAE has occurred.

Usage


respiratory_vae_tbl(.data)

Arguments

.data
The data passed in from respiratory_vae_query()

Details

Returns a tibble

Expects respiratory_vae_query() as the data argument

Makes the calculations according to the CDC NHSN VAE Calculator https://nhsn.cdc.gov/VAECalculator/vaecalc_v7.html

Value

A tibble

Author(s)

Steven P. Sanderson II, MPH

See Also

This one works better, more concise and works in 3.6.x and 4.0.x https://stackoverflow.com/questions/26553638/calculate-elapsed-time-since-last-event

VAE Calculator (requires javascript) https://nhsn.cdc.gov/VAECalculator/vaecalc_v7.html

Adding data.table to Depends: issue https://stackoverflow.com/questions/27980835/r-data-table-works-in-direct-call-but-same-function-in-a-package-fails

Examples


## Not run: 
respiratory_vae_query() %>%
  respiratory_vae_tbl()

respiratory_vae_query() %>%
  respiratory_vae_tbl() %>%
  save_to_excel(.file_name = "respiratory_vae")

## End(Not run)


[Package LICHospitalR version 0.2.0 Index]