save_to_excel {LICHospitalR}	R Documentation

Save a file to Excel

Description

Save a tibble/data.frame to an excel .xlsx file. The file will automatically with a save_dtime in the format of 20201109_132416 for November 11th, 2020 at 1:24:16PM.

Usage


save_to_excel(.data, .file_name)

Arguments

.data
The tibble/data.frame that you want to save as an .xlsx file.

.file_name
the name you want to give to the file.

Details

Requires a tibble/data.frame to be passed to it.

Value

A saved excel file

Author(s)

Steven P. Sanderson II, MPH

Examples


## Not run: 
coded_consults_query(.resp_pty = "013789") %>%
  save_to_excel(.file_name = "coded_consults")

coded_consults_query(.resp_pty = "013789") %>%
  coded_consults_top_providers_tbl(
  .top_n = 10
  , .attending_col = attending_md
  , .consultant_col = consultant
  ) %>%
  save_to_excel(.file_name = "coded_consults")

## End(Not run)

[Package LICHospitalR version 0.2.0 Index]