sql_left {LICHospitalR}	R Documentation

Use SQL LEFT type function in R

Description

Perform a SQL LEFT() on a piece of data

Usage


sql_left(text, num_char)

Arguments

text
A piece of text or data you want to take a left part of

num_char
How many characters of the text you want to take

Details

Returns a substring of text from the left side

Value

A left sided substring of data

Author(s)

Steven P Sanderson II, MPH

Examples


sql_left("This is some text", 4)


[Package LICHospitalR version 0.2.0 Index]