sql_mid {LICHospitalR}	R Documentation

Use SQL SUBSTRING() or Excel MID() type function in R

Description

Perform a SQL SUBSTRING() or Excel MID() type function on a piece of data

Usage


sql_mid(text, start_num, num_char)

Arguments

text
A piece of text or data you want to take a substring/mid of

start_num
Where you want to start your manipulation

num_char
How many characters o fthe text you want to take

Details

Returns a substring of text from the substring/mid point chosen

Value

A substring/mid section of text

Author(s)

Steven P Sanderson II, MPH

Examples


sql_mid("this is some text", 6, 2)


[Package LICHospitalR version 0.2.0 Index]