ts_alos_elos_tbl {LICHospitalR}	R Documentation

Time Series - ALOS/ELOS Excess Tibble

Description

This takes data from the ts_alos_elos_query() and makes an internal summary table using timetk::summarise_by_time with the following possible choices: "year", "month", "week", these are checked inside of the function, if something else is chose an error will be thrown and the function will exit. It defaults to "month".

Usage


ts_alos_elos_tbl(.data, .by_time = "month", .date_col)

Arguments

.data
The data passed in from ts_alos_elos_query()

.by_time
The choices are "year", "month", "week", defaults to "month"

.date_col
The column containing the date variable of interest

Details

Returns a tibble

Expects ts_alos_elos_query() as the data argument

Value

A tibble with the following columns:

date_col

visit_count

sum_days

sum_exp_days

alos

elos

excess_days

avg_excess

Author(s)

Steven P. Sanderson II, MPH

Examples


## Not run: 
library(healthyR)

data <- ts_alos_elos_query()

data %>%
  ts_alos_elos_tbl(.by_time = "year")

data %>%
  ts_alos_elos_tbl(.by_time = "month") %>%
  ts_plt(.date_col = dsch_date, .value_col = avg_excess)

data %>%
  ts_alos_elos_tbl()

data %>%
  ts_alos_elos_tbl(.by_time = "week") %>%
  save_to_excel(.file_name = "weekly_alos_excess")

## End(Not run)


[Package LICHospitalR version 0.2.0 Index]