ts_ed_arrivals_query {LICHospitalR}	R Documentation

Time Series - ER Arrivals Query

Description

This query gets the data necessary to do a time series analysis on emergency room arrivals.

Usage


ts_ed_arrivals_query()

Details

Requires a connection to DSS, uses both db_connect() and db_disconnect()

The start date is set to 2010-01-01

The end date is set dynamically to the end of the previous month.

It gets data from the WellSoft reporting server

The data comes back aggregated by arrival which is granular to the minute. The following columns come back:

date_col - date is of datetime, so down to the second

value - where value simply equals 1

Value

A tibble

Author(s)

Steven P. Sanderson II, MPH

Examples


## Not run: 
ts_ed_arrivals_query()

## End(Not run)


[Package LICHospitalR version 0.2.0 Index]