ts_ip_census_los_daily_query {LICHospitalR}	R Documentation

Time Series - Inpatient Census/LOS by Day

Description

Sometimes it is important to know what the census was on any given day, or what the average length of stay is on any given day, including for those patients that are not yet discharged. This can be easily achieved.

Usage


ts_ip_census_los_daily_query()

Details

Requires a connection to DSS, uses both db_connect() and db_disconnect()

The start date is set to 2001-01-01

The end date is not set in order to capture those that are still here.

It gets data from smsdss.BMH_PLM_PtAcct_V

Value

A tibble object

Author(s)

Steven P. Sanderson II, MPH

Examples


## Not run: 
ts_ip_census_los_daily_query()

## End(Not run)


[Package LICHospitalR version 0.2.0 Index]