ts_ip_census_los_daily_tbl {LICHospitalR}	R Documentation

Time Series - Inpatient Census/LOS by Day Tibble

Description

Sometimes it is important to know what the census was on any given day, or what the average length of stay is on any given day, including for those patients that are not yet discharged. This can be easily achieved. This will return one record for every account so the data will still need to be summarised.

For those accounts that are not yet discharged the date column that returns will be set to today (the day the function is run.)

This function can take a little bit of time to run while the join comparison runs.

Usage


ts_ip_census_los_daily_tbl(.data, .keep_nulls_only = FALSE)

Arguments

.data
The data passed from from ts_ip_census_los_daily_query()

.keep_nulls_only
A boolean that will keep only those records that have a NULL discharge date, meaning the patient is currently admitted. The default is FALSE which brings back all records.

Details

Requires the data from the ts_ip_census_los_daily_query()

Takes a single boolean parameter

Value

A tibble object

Author(s)

Steven P. Sanderson II, MPH

Examples


## Not run: 
ts_ip_census_los_daily_query() %>%
  ts_ip_census_los_daily_tbl()

ts_ip_census_los_daily_query() %>%
  ts_ip_census_los_daily_tbl(.keep_nulls_only = TRUE)

## End(Not run)


[Package LICHospitalR version 0.2.0 Index]