ts_ip_discharges_tbl {LICHospitalR}	R Documentation

Time Series - Inpatient Discharges Tibble

Description

This takes data from the ts_ip_discharges_query() and makes an internal summary table using timetk::summarise_by_time with the following possible choices: "year", "month", "week", and "day", these are checked inside of the function, if something else is chose an error will be thrown and the function will exit. It defaults to "month". This function also uses the timetk::pad_by_time function with a value of 0 for time periods that have no data.

Usage


ts_ip_discharges_tbl(.data, .by_time = "month", .date_col)

Arguments

.data
The data passed in from ts_ip_discharges_query()

.by_time
The choices are "year", "month", "week", "day" defaults to "month"

.date_col
The column containing the date variable of interest

Details

Returns a tibble

Expects ts_ip_discharges_query() as the data argument

Value

A tibble with the following columns:

date_col

value

Author(s)

Steven P. Sanderson II, MPH

Examples


## Not run: 
library(healthyR)

data <- ts_ip_discharges_query()

data %>%
  ts_ip_discharges_tbl(.by_time = "year")

data %>%
  ts_ip_discharges_tbl(.by_time = "month") %>%
  ts_plt(.date_col = dsch_date, .value_col = value)

data %>%
  ts_ip_discharges_tbl()

data %>%
  ts_ip_discharges_tbl(.by_time = "week") %>%
  save_to_excel(.file_name = "ip_discharges")

## End(Not run)


[Package LICHospitalR version 0.2.0 Index]