ts_median_excess_plt {LICHospitalR}	R Documentation

Create a plot showing the excess +/- of the median value

Description

Plot out the excess +/- of the median value grouped by certain time parameters

Usage


ts_median_excess_plt(
  .data,
  .date_col,
  .value_col,
  .x_axis,
  .ggplot_group_var,
  .years_back
)

Arguments

.data
The data that is being analyzed, data must be a tibble/data.frame

.date_col
The column of the tibble that holds the date

.value_col
The column that holds the value of interest

.x_axis
What is the be the x-axis, day, week, etc.

.ggplot_group_var
The variable to group the ggplot on

.years_back
How many yeas back do you want to go in order to compute the median value

Details

Supply data that you want to view and you will see the excess +/- of the median values over a specified time series tibble

Value

A ggplot2 plot

Author(s)

Steven P Sanderson II, MPH

Examples


library(timetk)
library(dplyr)
library(lubridate)

ts_ymwdh_tbl(
  .data = m4_daily
  , .date_col = date
) %>%
ts_median_excess_plt(
  .date_col = date
  , .value_col = value
  , .x_axis = wk
  , .ggplot_group_var = yr
  , .years_back = 5
)


[Package LICHospitalR version 0.2.0 Index]