ts_op_visits_query {LICHospitalR}	R Documentation

Time Series - Outpatient Visits Query

Description

This query gets the data necessary to do a time series analysis on outpatient visits.

Usage


ts_op_visits_query()

Details

Requires a connection to DSS, uses both db_connect() and db_disconnect()

The start date is set to 2001-01-01

The end date is set dynamically to the end of the previous month.

It gets data from smsdss.BMH_PLM_PtAcct_V

The data comes back aggregated by day, using the discharge date which is the same as the arrival date. The columns that are returned are:

date_col - date is of date

value - where value count of visits that day

This query excludes unitized and ED visits and those that start with 9

Value

A tibble

Author(s)

Steven P. Sanderson II, MPH

Examples


## Not run: 
ts_op_visits_query()

## End(Not run)


[Package LICHospitalR version 0.2.0 Index]