ts_readmit_excess_query {LICHospitalR}	R Documentation

Time Series - Readmission Excess Query

Description

This query gets the data necessary to do a time series analysis on the excess readmission rates.

Usage


ts_readmit_excess_query()

Details

Requires a connection to DSS, uses both db_connect() and db_disconnect()

The start date is set to 2016-04-01

The end date is set dynamically to the end of the previous month.

It gets data from:

smsdss.BMH_PLM_PtAcct_V

smsdss.pract_dim_v

Customer.Custom_DRG

smsdss.c_LIHN_Svc_Line_Tbl

smsdss.vReadmits

smsdss.c_Readmit_Dashboard_Bench_Tbl

smsdss.c_ppr_apr_drg_global_exclusions

Value

A tibble

Author(s)

Steven P. Sanderson II, MPH

Examples


## Not run: 
ts_readmit_excess_query()

## End(Not run)


[Package LICHospitalR version 0.2.0 Index]