weekly_psy_discharges_automation {LICHospitalR}	R Documentation

Weekly PSY Discharges Automation

Description

This takes in data directly from the weekly_psy_discharges_query() as the data argument

Usage


weekly_psy_discharges_automation(.data, .delete_file = FALSE, .email)

Arguments

.data
The data from myhealth_monthly_surgery_tbl()

.delete_file
FALSE is the default, so the file will be kept, TRUE will delete the file after function completion

.email
Provide the email address for the recipient. The email must be in double quotes like so: c("person@licommunityhospital.org;person2@licommunityhospital.org") using a semi-colon if there is more than one address.

Details

Will automatically keep the file unless otherwise specified

Sends file via email in Outlook

Author(s)

Steven P. Sanderson II, MPH

Examples


## Not run: 
library(RDCOMClient)
library(tidyverse)
weekly_psy_discharges_query() %>%
  weekly_psy_discharges_automation(.email = "someone@email.com")

## End(Not run)


[Package LICHospitalR version 0.2.0 Index]